+++ drivers/gpu/drm/panel/panel-tdo-y17p.c
@@ -218,7 +218,7 @@ static int tdo_y17p_probe(struct spi_device *spi)
 	if (!id)
 		return -ENODEV;
 
-	ctx->bus_format = (u32)id->data;
+	ctx->bus_format = (u32)(uintptr_t)id->data;
 
 	spi_set_drvdata(spi, ctx);
 	ctx->spi = spi;
